Hook up decorations; move EnvironmentDecorations to separate file.
