+++ core/frontend/src/DrawingViewState.ts
@@ -26,7 +26,7 @@ import { DisclosedTileTreeSet, TileGraphicType } from "./tile/internal";
 import { SceneContext } from "./ViewContext";
 import { OffScreenViewport } from "./Viewport";
 import { ViewRect } from "./ViewRect";
-import { ExtentLimits, ViewState2d, ViewState3d } from "./ViewState";
+import { AttachToViewportArgs, ExtentLimits, ViewState2d, ViewState3d } from "./ViewState";
 
 /** Strictly for testing.
  * @internal
@@ -319,8 +319,8 @@ export class DrawingViewState extends ViewState2d {
   }
 
   /** @internal */
-  public override attachToViewport(): void {
-    super.attachToViewport();
+  public override attachToViewport(args: AttachToViewportArgs): void {
+    super.attachToViewport(args);
     assert(undefined === this._attachment);
     this._attachment = this._attachmentInfo.createAttachment();
   }
