+++ core/frontend/src/ViewState.ts
@@ -6,15 +6,15 @@
  * @module Views
  */
 
-import { assert, BeEvent, Id64, Id64Arg, Id64String, JsonUtils } from "@itwin/core-bentley";
+import { assert, BeEvent, dispose, Id64, Id64Arg, Id64String, JsonUtils } from "@itwin/core-bentley";
 import {
-  Angle, AxisOrder, ClipVector, Constant, Geometry, IndexedPolyface, LongitudeLatitudeNumber, LowAndHighXY, LowAndHighXYZ, Map4d, Matrix3d,
-  Plane3dByOriginAndUnitNormal, Point2d, Point3d, PolyfaceBuilder, Range2d, Range3d, Ray3d, StrokeOptions, Transform, Vector2d, Vector3d, XAndY,
+  Angle, AxisOrder, ClipVector, Constant, Geometry, LongitudeLatitudeNumber, LowAndHighXY, LowAndHighXYZ, Map4d, Matrix3d,
+  Plane3dByOriginAndUnitNormal, Point2d, Point3d, Range2d, Range3d, Ray3d, Transform, Vector2d, Vector3d, XAndY,
   XYAndZ, XYZ, YawPitchRollAngles,
 } from "@itwin/core-geometry";
 import {
-  AnalysisStyle, AxisAlignedBox3d, Camera, Cartographic, ColorDef, FeatureAppearance, Frustum, GlobeMode, Gradient, GraphicParams, GridOrientationType,
-  ModelClipGroups, Npc, RenderMaterial, RenderSchedule, RenderTexture, SkyCube, SkySphere, SubCategoryOverride, TextureImageSpec, TextureMapping,
+  AnalysisStyle, AxisAlignedBox3d, Camera, Cartographic, ColorDef, FeatureAppearance, Frustum, GlobeMode, GridOrientationType,
+  ModelClipGroups, Npc, RenderSchedule, SubCategoryOverride,
   ViewDefinition2dProps, ViewDefinition3dProps, ViewDefinitionProps, ViewDetails, ViewDetails3d, ViewFlags, ViewStateProps,
 } from "@itwin/core-common";
 import { AuxCoordSystem2dState, AuxCoordSystem3dState, AuxCoordSystemState } from "./AuxCoordSys";
@@ -27,10 +27,8 @@ import { IModelApp } from "./IModelApp";
 import { IModelConnection } from "./IModelConnection";
 import { GeometricModel2dState, GeometricModelState } from "./ModelState";
 import { NotifyMessageDetails, OutputMessagePriority } from "./NotificationManager";
-import { GraphicType } from "./render/GraphicBuilder";
 import { RenderClipVolume } from "./render/RenderClipVolume";
 import { RenderMemory } from "./render/RenderMemory";
-import { RenderGraphicOwner } from "./render/RenderGraphic";
 import { RenderScheduleState } from "./RenderScheduleState";
 import { SheetViewState } from "./SheetViewState";
 import { SpatialViewState } from "./SpatialViewState";
@@ -43,9 +41,7 @@ import { ViewingSpace } from "./ViewingSpace";
 import { Viewport } from "./Viewport";
 import { ViewPose, ViewPose2d, ViewPose3d } from "./ViewPose";
 import { ViewStatus } from "./ViewStatus";
-import { Environment } from "@itwin/core-common";
-import { SkyBox } from "./render/RenderSystem";
-import { tryImageElementFromUrl } from "./ImageUtil";
+import { EnvironmentDecorations } from "./EnvironmentDecorations";
 
 /** Describes the largest and smallest values allowed for the extents of a [[ViewState]].
  * Attempts to exceed these limits in any dimension will fail, preserving the previous extents.
@@ -156,6 +152,11 @@ const scratchRay = Ray3d.createZero();
 const unitRange2d = Range2d.createXYXY(0, 0, 1, 1);
 const scratchRange2d = Range2d.createNull();
 const scratchRange2dIntersect = Range2d.createNull();
+
+/** @internal */
+export interface AttachToViewportArgs {
+  invalidateDecorations: () => void;
+}
 
 /** The front-end state of a [[ViewDefinition]] element.
  * A ViewState is typically associated with a [[Viewport]] to display the contents of the view on the screen. A ViewState being displayed by a Viewport is considered to be
@@ -1222,7 +1223,7 @@ export abstract class ViewState extends ElementState {
    * @see [[detachFromViewport]] from the inverse operation.
    * @internal
    */
-  public attachToViewport(): void {
+  public attachToViewport(_args: AttachToViewportArgs): void {
     if (this.isAttachedToViewport)
       throw new Error("Attempting to attach a ViewState that is already attached to a Viewport");
 
@@ -1272,235 +1273,6 @@ export abstract class ViewState extends ElementState {
   }
 }
 
-interface GroundPlaneDecorations {
-  readonly aboveParams: GraphicParams;
-  readonly belowParams: GraphicParams;
-}
-
-interface SkyBoxDecorations {
-  graphic?: RenderGraphicOwner;
-  promise?: Promise<SkyBox.CreateParams | undefined>;
-}
-
-class EnvironmentDecorations {
-  private readonly _view: ViewState3d;
-  private readonly _onLoaded: () => void;
-  private _environment: Environment;
-  private _ground?: GroundPlaneDecorations;
-  private _sky: SkyBoxDecorations;
-
-  public constructor(view: ViewState3d, onLoaded: () => void) {
-    this._environment = view.displayStyle.environment;
-    this._view = view;
-    this._onLoaded = onLoaded;
-
-    this._sky = { };
-    this.loadSky();
-    if (this._environment.displayGround)
-      this.loadGround();
-  }
-
-  public dispose(): void {
-    this._ground = undefined;
-
-    this._sky.graphic?.disposeGraphic();
-    this._sky.promise = undefined;
-  }
-
-  public setEnvironment(env: Environment): void {
-    const prev = this._environment;
-    if (prev === env)
-      return;
-
-    // Update ground plane
-    if (!env.displayGround || env.ground !== prev.ground)
-      this._ground = undefined;
-
-    if (env.displayGround && !this._ground)
-      this.loadGround();
-
-    // Update sky box
-    if (env.sky !== prev.sky)
-      this.loadSky();
-  }
-
-  public decorate(context: DecorateContext): void {
-    const env = this._environment;
-    if (env.displaySky && this._sky.graphic)
-      context.setSkyBox(this._sky.graphic);
-
-    if (!env.displayGround || !this._ground)
-      return;
-
-    const extents = this._view.getGroundExtents(context.viewport);
-    if (extents.isNull)
-      return;
-
-    const points: Point3d[] = [extents.low.clone(), extents.low.clone(), extents.high.clone(), extents.high.clone()];
-    points[1].x = extents.high.x;
-    points[3].x = extents.low.x;
-
-    const aboveGround = this._view.isEyePointAbove(extents.low.z);
-    const params = aboveGround ? this._ground.aboveParams : this._ground.belowParams;
-    const builder = context.createGraphicBuilder(GraphicType.WorldDecoration);
-    builder.activateGraphicParams(params);
-
-    const strokeOptions = new StrokeOptions();
-    strokeOptions.needParams = true;
-    const polyfaceBuilder = PolyfaceBuilder.create(strokeOptions);
-    polyfaceBuilder.toggleReversedFacetFlag();
-    const uvParams: Point2d[] = [Point2d.create(0, 0), Point2d.create(1, 0), Point2d.create(1, 1), Point2d.create(0, 1)];
-    polyfaceBuilder.addQuadFacet(points, uvParams);
-    const polyface = polyfaceBuilder.claimPolyface(false);
-
-    builder.addPolyface(polyface, true);
-    context.addDecorationFromBuilder(builder);
-  }
-
-  private loadGround(): void {
-    assert(undefined === this._ground);
-    const aboveParams = this.createGroundParams(true);
-    const belowParams = this.createGroundParams(false);
-    if (aboveParams && belowParams)
-      this._ground = { aboveParams, belowParams };
-  }
-
-  private createGroundParams(above: boolean): GraphicParams | undefined {
-    // Create a gradient texture.
-    const ground = this._environment.ground;
-    const values = [0, 0.25, 0.5 ];
-    const color = above ? ground.aboveColor : ground.belowColor;
-    const alpha = above ? 0x80 : 0x85;
-    const groundColors = [color.withTransparency(0xff), color, color];
-    groundColors[1] = groundColors[2] = color.withTransparency(alpha);
-
-    const gradient = new Gradient.Symb();
-    gradient.mode = Gradient.Mode.Spherical;
-    gradient.keys = [{ color: groundColors[0], value: values[0] }, { color: groundColors[1], value: values[1] }, { color: groundColors[2], value: values[2] }];
-    const texture = IModelApp.renderSystem.getGradientTexture(gradient, this._view.iModel);
-    if (!texture)
-      return undefined;
-
-    // Create a material using the gradient texture
-    const matParams = new RenderMaterial.Params();
-    matParams.diffuseColor = ColorDef.white;
-    matParams.shadows = false;
-    matParams.ambient = 1;
-    matParams.diffuse = 0;
-
-    const mapParams = new TextureMapping.Params();
-    const transform = new TextureMapping.Trans2x3(0, 1, 0, 1, 0, 0);
-    mapParams.textureMatrix = transform;
-    matParams.textureMapping = new TextureMapping(texture, mapParams);
-    const material = IModelApp.renderSystem.createMaterial(matParams, this._view.iModel);
-    if (!material)
-      return undefined;
-
-    // Create GraphicParams using the material.
-    const params = new GraphicParams();
-    params.lineColor = gradient.keys[0].color;
-    params.fillColor = ColorDef.white;  // Fill should be set to opaque white for gradient texture...
-    params.material = material;
-
-    return params;
-  }
-
-  private loadSky(): void {
-    const promise = this.loadSkyParams();
-    this._sky.promise = promise;
-
-    promise.then((params) => {
-      if (promise === this._sky.promise)
-        this.setSky(params ?? this.createSkyGradientParams());
-    });
-
-    promise.catch(() => {
-      if (this._sky.promise === promise)
-        this.setSky(this.createSkyGradientParams());
-    });
-  }
-
-  private setSky(params: SkyBox.CreateParams): void {
-    this._sky.promise = undefined;
-    const gf = IModelApp.renderSystem.createSkyBox(params);
-    this._sky.graphic = gf ? IModelApp.renderSystem.createGraphicOwner(gf) : undefined;
-    this._onLoaded();
-  }
-
-  private async loadSkyParams(): Promise<SkyBox.CreateParams | undefined> {
-    const sky = this._environment.sky;
-    if (sky instanceof SkyCube) {
-      const key = this.createCubeImageKey(sky);
-      const existingTexture = IModelApp.renderSystem.findTexture(key, this._view.iModel);
-      if (existingTexture)
-        return SkyBox.CreateParams.createForCube(existingTexture);
-
-      // Some faces may use the same image. Only request each image once.
-      const specs = new Set<string>([sky.images.front, sky.images.back, sky.images.left, sky.images.right, sky.images.top, sky.images.bottom]);
-      const promises = [];
-      for (const spec of specs)
-        promises.push(this.imageFromSpec(spec));
-
-      return Promise.all(promises).then((images) => {
-        const idToImage = new Map<TextureImageSpec, HTMLImageElement>();
-        let index = 0;
-        for (const spec of specs) {
-          const image = images[index++];
-          if (!image)
-            return undefined;
-          else
-            idToImage.set(spec, image);
-        }
-
-        // eslint-disable-next-line deprecation/deprecation
-        const params = new RenderTexture.Params(key, RenderTexture.Type.SkyBox);
-        const txImgs = [
-          idToImage.get(sky.images.front)!, idToImage.get(sky.images.back)!, idToImage.get(sky.images.top)!,
-          idToImage.get(sky.images.bottom)!, idToImage.get(sky.images.right)!, idToImage.get(sky.images.left)!,
-        ];
-
-        const texture = IModelApp.renderSystem.createTextureFromCubeImages(txImgs[0], txImgs[1], txImgs[2], txImgs[3], txImgs[4], txImgs[5], this._view.iModel, params);
-        return texture ? SkyBox.CreateParams.createForCube(texture) : undefined;
-      });
-    } else if (sky instanceof SkySphere) {
-      const rotation = 0; // ###TODO where is this supposed to come from?
-      let texture = IModelApp.renderSystem.findTexture(sky.image, this._view.iModel);
-      if (!texture) {
-        const image = await this.imageFromSpec(sky.image);
-        if (image) {
-          texture = IModelApp.renderSystem.createTexture({
-            image: { source: image },
-            ownership: { iModel: this._view.iModel, key: sky.image },
-          });
-        }
-      }
-
-      if (!texture)
-        return undefined;
-
-      return SkyBox.CreateParams.createForSphere(new SkyBox.SphereParams(texture, rotation), this._view.iModel.globalOrigin.z);
-    } else {
-      return this.createSkyGradientParams();
-    }
-  }
-
-  private createCubeImageKey(sky: SkyCube): string {
-    const i = sky.images;
-    return `skycube:${i.front}:${i.back}:${i.left}:${i.right}:${i.top}:${i.bottom}`;
-  }
-
-  private createSkyGradientParams(): SkyBox.CreateParams {
-    return SkyBox.CreateParams.createForGradient(this._environment.sky.gradient, this._view.iModel.globalOrigin.z);
-  }
-
-  private async imageFromSpec(spec: TextureImageSpec): Promise<HTMLImageElement | undefined> {
-    if (Id64.isValidId64(spec))
-      return (await IModelApp.renderSystem.loadTextureImage(spec, this._view.iModel))?.image;
-
-    return tryImageElementFromUrl(spec);
-  }
-}
-
 /** Defines the state of a view of 3d models.
  * @see [ViewState Parameters]($docs/learning/frontend/views#viewstate-parameters)
  * @public
@@ -1508,6 +1280,7 @@ class EnvironmentDecorations {
 export abstract class ViewState3d extends ViewState {
   private readonly _details: ViewDetails3d;
   private readonly _modelClips: Array<RenderClipVolume | undefined> = [];
+  private _environmentDecorations?: EnvironmentDecorations;
   /** @internal */
   public static override get className() { return "ViewDefinition3d"; }
   /** True if the camera is valid. */
@@ -2215,23 +1988,9 @@ export abstract class ViewState3d extends ViewState {
 
   public override decorate(context: DecorateContext): void {
     super.decorate(context);
-    // ###TODO this.drawSkyBox(context);
-    this.drawGroundPlane(context);
+    if (this._environmentDecorations)
+      this._environmentDecorations.decorate(context);
   }
-
-  /** @internal */
-  // ###TODO protected drawSkyBox(_context: DecorateContext): void {
-  // ###TODO   const style3d = this.getDisplayStyle3d();
-  // ###TODO   if (!style3d.environment.displaySky)
-  // ###TODO     return;
-
-  // ###TODO   const vp = context.viewport;
-  // ###TODO   const skyBoxParams = style3d.loadSkyBoxParams(vp.target.renderSystem, vp);
-  // ###TODO   if (undefined !== skyBoxParams) {
-  // ###TODO     const skyBoxGraphic = IModelApp.renderSystem.createSkyBox(skyBoxParams);
-  // ###TODO     context.setSkyBox(skyBoxGraphic!);
-  // ###TODO   }
-  // ###TODO }
 
   /** Returns the ground elevation taken from the environment added with the global z position of this imodel. */
   public getGroundElevation(): number {
@@ -2275,62 +2034,6 @@ export abstract class ViewState3d extends ViewState {
     extents.high.addScaledInPlace(Vector3d.create(1, 1, 1), radius);
     extents.low.z = extents.high.z = elevation;
     return extents;
-  }
-
-  /** @internal */
-  protected drawGroundPlane(context: DecorateContext): void {
-    const extents = this.getGroundExtents(context.viewport);
-    if (extents.isNull)
-      return;
-
-    if (!this.getDisplayStyle3d().environment.displayGround)
-      return;
-
-    const ground = this.getDisplayStyle3d().environment.ground;
-    const points: Point3d[] = [extents.low.clone(), extents.low.clone(), extents.high.clone(), extents.high.clone()];
-    points[1].x = extents.high.x;
-    points[3].x = extents.low.x;
-
-    const aboveGround = this.isEyePointAbove(extents.low.z);
-    const gradient = new Gradient.Symb(); // ###TODO ground.getGroundPlaneGradient(aboveGround);
-    const texture = context.viewport.target.renderSystem.getGradientTexture(gradient, this.iModel);
-    if (!texture)
-      return;
-
-    const matParams = new RenderMaterial.Params();
-    matParams.diffuseColor = ColorDef.white;
-    matParams.shadows = false;
-    matParams.ambient = 1;
-    matParams.diffuse = 0;
-
-    const mapParams = new TextureMapping.Params();
-    const transform = new TextureMapping.Trans2x3(0, 1, 0, 1, 0, 0);
-    mapParams.textureMatrix = transform;
-    matParams.textureMapping = new TextureMapping(texture, mapParams);
-    const material = context.viewport.target.renderSystem.createMaterial(matParams, this.iModel);
-    if (!material)
-      return;
-
-    const params = new GraphicParams();
-    params.lineColor = gradient.keys[0].color;
-    params.fillColor = ColorDef.white;  // Fill should be set to opaque white for gradient texture...
-    params.material = material;
-
-    const builder = context.createGraphicBuilder(GraphicType.WorldDecoration);
-    builder.activateGraphicParams(params);
-
-    /// ### TODO: Until we have more support in geometry package for tracking UV coordinates of higher level geometry
-    // we will use a PolyfaceBuilder here to add the ground plane as a quad, claim the polyface, and then send that to the GraphicBuilder
-    const strokeOptions = new StrokeOptions();
-    strokeOptions.needParams = true;
-    const polyfaceBuilder = PolyfaceBuilder.create(strokeOptions);
-    polyfaceBuilder.toggleReversedFacetFlag();
-    const uvParams: Point2d[] = [Point2d.create(0, 0), Point2d.create(1, 0), Point2d.create(1, 1), Point2d.create(0, 1)];
-    polyfaceBuilder.addQuadFacet(points, uvParams);
-    const polyface = polyfaceBuilder.claimPolyface(false);
-
-    builder.addPolyface(polyface, true);
-    context.addDecorationFromBuilder(builder);
   }
 
   /** @internal */
@@ -2423,6 +2126,20 @@ export abstract class ViewState3d extends ViewState {
 
     frustum.multiply(transitionTransform);
     return this.setupFromFrustum(frustum);
+  }
+
+  public override attachToViewport(args: AttachToViewportArgs): void {
+    super.attachToViewport(args);
+
+    const removeListener = this.displayStyle.settings.onEnvironmentChanged.addListener((env) => {
+      this._environmentDecorations?.setEnvironment(env);
+    });
+
+    this._environmentDecorations = new EnvironmentDecorations(this, () => args.invalidateDecorations(), () => removeListener());
+  }
+
+  public override detachFromViewport(): void {
+    this._environmentDecorations = dispose(this._environmentDecorations);
   }
 }
